Updated the attachments partial
Had to change cmfive FILE_ROOT var, as it points at the wrong place in systems that symlink the system folder only. So uploads were going to the wrong place
It turns out the thumbnail generator lib is missing completely! I spent a fair bit of time when doing the partial trying to work out why images weren’t showing up, turned out to be multiple things but we’ll need to work out what to do