Merge conflict destroyed DbObject, thats fixed
Moved the templates and public files (css, js, templates) to system, where the user can override in their project root with the same layout if they desire.
This means that when we push changed to the layout that are important (i.e. when we move to Foundation) then everyone who has this project will get an update with the new layout but could override with their own layout (or copy the current layout for legacy reasons)