Refactor install guide
Moved the ever-growing index.php install file that handled everything
into split files, each responsible for a unique area of the install. The
idea for this will be that the config.php in assets will be the final config
by the end of the install. But for now nothing works well, it's writing the
config straight away to the project root...