Fix Web caching, add self healing PDO
Merging 0.8.3 into TaskChanges presented multiple problems.
Mainly the fact that because the config was cached prior to merging
I couldn't get the Timelog service method. So I made Web regenerate
its caches when it doesn't find a required service. This means, in the
spirit of being a 'mature' framework, if you misspell a config module,
it will keep regenerating until you fix it, which you'll need to do anyway
for the module to work. On this, I also added a small install function
for our DbPDO class to check if the missing table name is also a module,
it then tries to run the install/db.sql if one exists. Also removed the invasive
update file but left it in db.sql as a comment, it replaces TaskTime with Timelog
and probably isn't best idea.