Added a multitude of functionality, see description
Added the Html namespace to the autoloader, so you don't need to require
a million html files manually

Added InputField helper classes like Button, Number, Text, Password etc.

Added the GlobalAttributes and Events traits for Html classes:
This is a big one, the Events trait already used in GlobalAttributes, so if
you want both, you only need to include GlobalAttributes.

They are traits defining attributes that are available to every HTML element,
I made them as traits because it helps splitting up responsibility of files and
enhances the usability of one or more, as the developer sees fit