Added counter to generic forms, also added method of counting classes on a page
This replaces the counter for comments and attachments that was on the task page.
For e.g: attachments all have the "attachment" class on them, so you add a span label
with the class "cmfive__count-attachment" and the js on the page will look for all "attachment"
classes, count them and put the result inside the span with the "cmfive__count-attachment"
class.