Fixed issue with integer being added into an array
When we're setting a singleton value, that is only one instance of the value would ever need
to exist, then we must use Config::set. Config::append is used when we know there may be
other values in the way and we want to be non-intrusive, i.e. we want to _append_ to the
Config item that we're setting. By default, Config::append will create the item in an array
if it doesn't yet exist under the assumption that you can have more than one value in that
place.